Split files and hooks
